Add files via upload
Was able to install a firewall and find unnecessary applications. Regarding adding a monitoring system phpmyadmin allows access to the databases but I am struggling to implement the monitoring with zabbix due to deprecated errors. So I'm unsure and could use some help on how I can implement the monitoring.